refactor: Phase 5 [5.5] migrate training/ — move train/* to training/runners/, add deprecation shims (BPR)
