+++ pipelines/train/cat_head.pipe.py
@@ -12,7 +12,7 @@ if _src not in sys.path:
 from configs.paths import IoPaths, RESULTS_ROOT, EmbeddingEngine
 from configs.experiment import ExperimentConfig
 from data.folds import FoldCreator, TaskType
-from train.category.cross_validation import run_cv
+from training.runners.category_cv import run_cv
 from common.ml_history import MLHistory, DatasetHistory
 from configs.globals import CATEGORIES_MAP
