+++ src/train/shared/evaluate.py
@@ -1,34 +1,14 @@
-from typing import Union
+"""Shim — canonical location is training.shared_evaluate (Phase 5).
 
-import numpy as np
-import torch
-from sklearn.metrics import classification_report
-from torch import nn
-from torch.utils.data import DataLoader
+This module will be removed at the end of Phase 6.
+"""
+import warnings as _warnings
 
+_warnings.warn(
+    "train.shared.evaluate is deprecated; use training.shared_evaluate instead. "
+    "This shim will be removed at the end of Phase 6.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-def evaluate(model: nn.Module, loader: DataLoader, device: torch.device,
-             best_state: dict = None
-             ) -> Union[str, dict]:
-    model.eval()
-    if best_state is not None:
-        model.load_state_dict(best_state)
-    preds_list, truths_list = [], []
-    with torch.no_grad():
-        for X_batch, y_batch in loader:
-            X_batch = X_batch.to(device, non_blocking=True)
-            logits = model(X_batch)
-            preds_list.append(logits.argmax(dim=1))
-            truths_list.append(y_batch)
-
-    # Single GPU→CPU transfer for all predictions
-    preds = torch.cat(preds_list).cpu().numpy()
-    truths = torch.cat(truths_list).numpy()
-
-    report = classification_report(
-        truths,
-        preds,
-        output_dict=True,
-        zero_division=0,
-    )
-    return report
+from training.shared_evaluate import evaluate  # noqa: F401
