+++ src/training/runners/next_eval.py
@@ -1,0 +1,1 @@
+from training.shared_evaluate import evaluate  # noqa: F401
